@@ -1,11 +1,8 @@
 module github.com/jjeffery/apigatewayproxy
 
-go 1.17
+go 1.24
 
 require (
-	github.com/aws/aws-lambda-go v1.27.1
+	github.com/aws/aws-lambda-go v1.48.0
 	github.com/davecgh/go-spew v1.1.1
-	github.com/jjeffery/kv v0.8.1
 )
-
-require github.com/stretchr/testify v1.7.0 // indirect
